Move config validation to utils
