+++ skybaks/cup_manager/templates/config.xml
@@ -1,4 +1,5 @@
 {% import "cup_manager/config_data.xml" as config_data %}
+{% from "cup_manager/icons.xml" import icons %}
 
 {% set sel_col = "aaa" %}
 {% set def_col = "000" %}
@@ -27,7 +28,18 @@
     {% for tab in config_tabs %}
       <label pos="{{ 10 + (20.5 * loop.index0) }} -3.5" size="20 7" textprefix="$o" text="{{ tab.name }}" textsize="1.3" valign="center2" halign="center" {{ set_selection_color(tab.selected) }} action="{{ id }}__config_tab_{{ loop.index0 }}"/>
     {% endfor %}
-      <quad pos="0 -7" size="218 1" halign="left" valign="top" bgcolor="{{ sel_col }}{{ sel_aph }}" />
+      <quad pos="0 -7" size="109 1" halign="left" valign="top" bgcolor="{{ sel_col }}{{ sel_aph }}" />
+  </frame>
+
+  {# toolbar buttons #}
+  <frame pos="0 50">
+    {{ config_data.button_icon(47, 1, 10, 10, icons.check, action=id + "__toolbar_validate", textsize=2.5, z_index=15) }}
+    <quad pos="48 0" size="30 8" halign="left" valign="top" bgcolor="{{ def_col }}{{ def_aph }}" z-index="-40" />
+    <label pos="58 -4" size="20 8" halign="left" valign="center2" text="Validate" textsize="2.5" z-index="15" />
+
+    {{ config_data.button_icon(78, 1, 10, 10, icons.save, action=id + "__toolbar_save", textsize=2.5, z_index=15) }}
+    <quad pos="79 0" size="30 8" halign="left" valign="top" bgcolor="{{ def_col }}{{ def_aph }}" z-index="-40" />
+    <label pos="89 -4" size="20 8" halign="left" valign="center2" text="Save" textsize="2.5" z-index="15" />
   </frame>
 
   {# sidebar #}
