+++ skybaks/cup_manager/templates/icons.xml
@@ -4,11 +4,13 @@
     exclamation_circle="",
     exclamation_triangle="",
     info_circle="",
+    kenny_save="",
     minus_circle="",
     pencil="",
     plus_circle="",
     question_mark="",
     question_mark_circle="",
+    save="",
     times_circle="",
     translation="",
     trash="",
